+++ test/code/test_renderPerfFromJson.cjs
